+++ langwatch/src/components/settings/ModelProviderScopeSection.tsx
@@ -1,12 +1,4 @@
-import {
-  Box,
-  createListCollection,
-  HStack,
-  Text,
-  VStack,
-} from "@chakra-ui/react";
-import { Building2, Folder, Users } from "lucide-react";
-import { useMemo } from "react";
+import { Text, VStack } from "@chakra-ui/react";
 import type {
   ModelProviderScopeType,
   ScopeSelection,
@@ -14,16 +6,9 @@ import type {
   UseModelProviderFormState,
 } from "../../hooks/useModelProviderForm";
 import type { MaybeStoredModelProvider } from "../../server/modelProviders/registry";
-import { Select } from "../ui/select";
 import { SmallLabel } from "../SmallLabel";
 import { ProviderScopeChips } from "./ProviderScopeChips";
-
-type ScopeOption = {
-  value: string;
-  label: string;
-  scopeType: ModelProviderScopeType;
-  scopeId: string;
-};
+import { ScopeChipPicker } from "./ScopeChipPicker";
 
 const SCOPE_DESCRIPTION_SINGLE: Record<ModelProviderScopeType, string> = {
   PROJECT: "Only this project can use this provider.",
@@ -56,28 +41,14 @@ function summariseSelection(scopes: ScopeSelection[]): string {
   return `Shared across ${parts.join(" + ")}.`;
 }
 
-const ScopeIcon = ({ scopeType }: { scopeType: ModelProviderScopeType }) => {
-  if (scopeType === "ORGANIZATION") return <Building2 size={16} aria-hidden />;
-  if (scopeType === "TEAM") return <Users size={16} aria-hidden />;
-  return <Folder size={16} aria-hidden />;
-};
-
 /**
- * Scope picker for model providers.
+ * Model-provider scope picker. For NEW providers, delegates to the shared
+ * `ScopeChipPicker` primitive (also consumed by the role-based default
+ * model lines). For EXISTING providers the section renders read-only
+ * chips — scope changes on a persisted credential happen by delete +
+ * recreate so we never silently re-parent a credential across orgs/teams.
  *
- * For NEW providers, renders a Chakra multi-select over the organization,
- * every team the caller is a member of, and the projects inside those
- * teams. Users pick one or more scopes; every selected entry becomes a
- * ModelProviderScope row on save. The service runs fail-closed authz per
- * entry — selecting a scope the caller cannot manage rejects the whole
- * write, there is no partial-success path.
- *
- * For EXISTING providers the section is read-only: scope changes on a
- * persisted credential happen by delete+recreate so we never silently
- * re-parent a credential across orgs/teams.
- *
- * For personal-account projects (no org/team context) the section
- * renders nothing.
+ * Personal-account projects (no org/team context) render nothing.
  */
 export function ProviderScopeSection({
   state,
@@ -101,76 +72,11 @@ export function ProviderScopeSection({
   organizationName?: string;
   projectId?: string;
   projectName?: string;
-  /** Teams the caller can pick for TEAM scope. Falls back to [{id:teamId}] when omitted. */
   availableTeams?: Array<{ id: string; name: string }>;
-  /** Projects the caller can pick for PROJECT scope. Falls back to [{id:projectId}]. */
   availableProjects?: Array<{ id: string; name: string; teamId?: string }>;
 }) {
   const isExisting = Boolean(provider.id);
   const hasOrgOrTeam = Boolean(organizationId ?? teamId);
-
-  // Build options from the accessible set. Falls back to the active
-  // context IDs when the caller hasn't provided a full team/project
-  // list — single-scope flows keep working unchanged.
-  const options = useMemo<ScopeOption[]>(() => {
-    const out: ScopeOption[] = [];
-    if (organizationId) {
-      out.push({
-        value: `ORGANIZATION:${organizationId}`,
-        label: organizationName ?? "Organization",
-        scopeType: "ORGANIZATION",
-        scopeId: organizationId,
-      });
-    }
-    const teams =
-      availableTeams && availableTeams.length > 0
-        ? availableTeams
-        : teamId
-          ? [{ id: teamId, name: teamName ?? "Team" }]
-          : [];
-    for (const team of teams) {
-      out.push({
-        value: `TEAM:${team.id}`,
-        label: team.name,
-        scopeType: "TEAM",
-        scopeId: team.id,
-      });
-    }
-    const projects =
-      availableProjects && availableProjects.length > 0
-        ? availableProjects
-        : projectId
-          ? [{ id: projectId, name: projectName ?? "Project" }]
-          : [];
-    for (const project of projects) {
-      out.push({
-        value: `PROJECT:${project.id}`,
-        label: project.name,
-        scopeType: "PROJECT",
-        scopeId: project.id,
-      });
-    }
-    return out;
-  }, [
-    organizationId,
-    organizationName,
-    teamId,
-    teamName,
-    projectId,
-    projectName,
-    availableTeams,
-    availableProjects,
-  ]);
-
-  const collection = useMemo(
-    () => createListCollection({ items: options }),
-    [options],
-  );
-
-  const selectedValues = useMemo(
-    () => state.scopes.map((s) => `${s.scopeType}:${s.scopeId}`),
-    [state.scopes],
-  );
 
   if (isExisting) {
     const storedScopes: ScopeSelection[] =
@@ -208,78 +114,17 @@ export function ProviderScopeSection({
   if (!hasOrgOrTeam) return null;
 
   return (
-    <VStack align="start" width="full" gap={2}>
-      <SmallLabel>Scope</SmallLabel>
-      <Select.Root
-        collection={collection}
-        value={selectedValues}
-        multiple
-        onValueChange={(details) => {
-          const picked = new Set(details.value);
-          const next = options
-            .filter((o) => picked.has(o.value))
-            .map((o) => ({ scopeType: o.scopeType, scopeId: o.scopeId }));
-          actions.setScopes(next);
-        }}
-      >
-        <Select.Trigger>
-          <Select.ValueText placeholder="Pick one or more scopes">
-            {() => {
-              if (state.scopes.length === 0) return "Pick one or more scopes";
-              return <ProviderScopeChips scopes={state.scopes} />;
-            }}
-          </Select.ValueText>
-        </Select.Trigger>
-        <Select.Content>
-          {options.some((o) => o.scopeType === "ORGANIZATION") && (
-            <Select.ItemGroup label="Organization">
-              {options
-                .filter((o) => o.scopeType === "ORGANIZATION")
-                .map((option) => (
-                  <Select.Item key={option.value} item={option}>
-                    <HStack gap={2}>
-                      <ScopeIcon scopeType="ORGANIZATION" />
-                      <Text>{option.label}</Text>
-                    </HStack>
-                  </Select.Item>
-                ))}
-            </Select.ItemGroup>
-          )}
-          {options.some((o) => o.scopeType === "TEAM") && (
-            <Select.ItemGroup label="Teams">
-              {options
-                .filter((o) => o.scopeType === "TEAM")
-                .map((option) => (
-                  <Select.Item key={option.value} item={option}>
-                    <HStack gap={2}>
-                      <ScopeIcon scopeType="TEAM" />
-                      <Text>{option.label}</Text>
-                    </HStack>
-                  </Select.Item>
-                ))}
-            </Select.ItemGroup>
-          )}
-          {options.some((o) => o.scopeType === "PROJECT") && (
-            <Select.ItemGroup label="Projects">
-              {options
-                .filter((o) => o.scopeType === "PROJECT")
-                .map((option) => (
-                  <Select.Item key={option.value} item={option}>
-                    <HStack gap={2}>
-                      <ScopeIcon scopeType="PROJECT" />
-                      <Text>{option.label}</Text>
-                    </HStack>
-                  </Select.Item>
-                ))}
-            </Select.ItemGroup>
-          )}
-        </Select.Content>
-      </Select.Root>
-      <Box>
-        <Text fontSize="xs" color="gray.600">
-          {summariseSelection(state.scopes)}
-        </Text>
-      </Box>
-    </VStack>
+    <ScopeChipPicker
+      value={state.scopes}
+      onChange={(next) => actions.setScopes(next)}
+      organizationId={organizationId}
+      organizationName={organizationName}
+      teamId={teamId}
+      teamName={teamName}
+      projectId={projectId}
+      projectName={projectName}
+      availableTeams={availableTeams}
+      availableProjects={availableProjects}
+    />
   );
 }
